1.add cmake support. \n. compile test tool when buiding current project. if not, only build the library. \n3. move ui_control to test/ cause it seems a test tool
